Move stock implementations into SoupStock.swift
